Move to a better connection routing which reuses the existing clients.  This is closer to the Client example in the NimBLE project and seems to be a bit faster and a bit more reliable.
The most common problem is that it just won't connect to the light.  I think this is a connection quality problem, not a code
problem.  I'm still trying to improve it.

The other problem is that you can't stay subscribed to the notifications and push writes at the same time.  That might
be me doing it wrong though.