Wow!  This is the most stable version yet!
With the hel of @h2zero the callbacks are now connected all the time,
reconnecting works more reliably, and when you're in range the initial
connection is fast and semi-reliable.

I might put some connect retries in next, but I think it's time
to work on testing in production now.

Some background here:   https://github.com/h2zero/NimBLE-Arduino/issues/316