Closing in on Version 1.0
It's working now.

It supports multiple ESP32s on one network, but you need to know the IP address of the one
you want to use to communicate with the lights. This could be made automatic by some kind of
election to see who has the best RSSI.  That's something for the next version though.

Next steps I neeed to tidy up the code a bit, update the readme and write some examples.

It will also be interesting to see if they can maintain a connection for days, or if that causes problems.
Perhaps an automatic disconnect after 10 mins is a good idea.  Also, if the connection is maintained then
the lights dont respond to scan, which make sense.